Crate auxtools

For when BYOND is not enough. Probably often.

Re-exports

pub use raw_types::variables::VariableNameIdTable;

pub use inventory;

pub use ctor;

Modules

debug

raw_types

sigscan

Macros

byond_ffi_fn

Creates a normal byond ffi function that can be called in DM with call.

byond_string

find_signature

find_signature_inner

find_signatures

pin_dll

The pin_dll! macro is used to determine whether the dll handle auxtools takes on Windows is pinned. For reference, a dll with a pinned handle cannot be unloaded during execution of the host process - termination of the host is the only way to unload the dll and release the lock on the corresponding file.

runtime

This macro makes instantiating Runtimes a (little bit) easier.

signature

signature_struct

signatures

universal_signature

version_dependent_signature

Structs

List

A wrapper around Values that make working with lists a little easier

Proc

Used to hook and call procs.

Runtime

Represents a byond runtime, sort of. This will probably drastically in the future.

StringRef

A wrapper around Values that make working with strings a little easier

Value

Value represents any value a DM variable can hold, such as numbers, strings, datums, etc.

WeakValue

A weak reference to some datum or atom in the game.

Constants

BYONDCORE

Statics

PIN_DLL

Type Definitions

DMResult

Attribute Macros

full_shutdown

hook

The hook attribute is used to define functions that may be used as proc hooks, and to optionally hook those procs upon library initialization.

init

runtime_handler

shutdown