Macro auxtools::byond_ffi_fn

macro_rules! byond_ffi_fn {
    ($name:ident() $body:block) => { ... };
    ($name:ident($($arg:ident),* $(, ...$rest:ident)?) $body:block) => { ... };
}

Creates a normal byond ffi function that can be called in DM with call.

You should favour hooks over these when working with auxtools.

Examples

// byond_ffi_fn! { my_proc(_input) {
//     Some("Hello, BYOND!".to_owned())
// }