Crate auxtools

For when BYOND is not enough. Probably often.

Re-exports

pub use raw_types::variables::VariableNameIdTable;

pub use inventory;

Modules

debug

raw_types

sigscan

Macros

byond_ffi_fn

Creates a normal byond ffi function that can be called in DM with call.

byond_string

runtime

This macro makes instantiating Runtimes a (little bit) easier.

signature

signatures

Structs

List

A wrapper around Values that make working with lists a little easier

Proc

Used to hook and call procs.

Runtime

Represents a byond runtime, sort of. This will probably drastically in the future.

StringRef

A wrapper around Values that make working with strings a little easier

Value

Value represents any value a DM variable can hold, such as numbers, strings, datums, etc.

WeakValue

A weak reference to some datum or atom in the game.

Constants

BYONDCORE

Type Definitions

DMResult

Attribute Macros

hook

The hook attribute is used to define functions that may be used as proc hooks, and to optionally hook those procs upon library initialization.

init

runtime_handler

shutdown