Struct auxtools::Value

pub struct Value {
    pub raw: Value,
    /* private fields */
}

Value represents any value a DM variable can hold, such as numbers, strings, datums, etc.

There’s a lot of lifetime shenanigans going on, the gist of it is to just not keep Values around for longer than your hook’s execution.

Fields
raw: Value
Implementations

impl Value

pub unsafe fn new(tag: ValueTag, data: ValueData) -> Value

Creates a new value from raw tag and data. Use if you know what you are doing.

pub fn globals() -> Value

Equivalent to DM’s global.vars.

pub fn world() -> Value

Equivalent to DM’s world.

pub fn null() -> Value

Equivalent to DM’s null.

pub unsafe fn turf_by_id_unchecked(id: u32) -> Value

Gets a turf by ID, without bounds checking. Use turf_by_id if you’re not sure about how to check the bounds.

pub fn turf_by_id(id: u32) -> DMResult

Gets a turf by ID, with bounds checking.

pub fn turf(x: u32, y: u32, z: u32) -> DMResult

Gets a turf by coordinates.

pub fn get<S: Into<StringRef>>(&self, name: S) -> DMResult

Gets a variable by name.

pub fn get_number<S: Into<StringRef>>(&self, name: S) -> DMResult<f32>

Gets a variable by name and safely casts it to a float.

pub fn get_string<S: Into<StringRef>>(&self, name: S) -> DMResult<String>

Gets a variable by name and safely casts it to a string.

pub fn get_list<S: Into<StringRef>>(&self, name: S) -> DMResult<List>

Gets a variable by name and safely casts it to a list::List.

pub fn set<S: Into<StringRef>, V: Into<Value>>(
&self,
name: S,
value: V
) -> DMResult<()>

Sets a variable by name to a given value.

pub fn as_number(&self) -> DMResult<f32>

Check if the current value is a number and casts it.

pub fn as_string(&self) -> DMResult<String>

Check if the current value is a string and casts it.

pub fn as_list(&self) -> DMResult<List>

Check if the current value is a list and casts it.

pub fn call<S: AsRef<str>>(&self, procname: S, args: &[&Value]) -> DMResult

Calls a method of the value with the given arguments.

Examples:

This example is equivalent to src.explode(3) in DM.

src.call("explode", &[&Value::from(3.0)]);

pub fn to_dmstring(&self) -> DMResult<StringRef>

pub fn to_string(&self) -> DMResult<String>

pub fn get_type(&self) -> Result<String, Runtime>

Gets the type of the Value as a string

pub fn is_exact_type<S: AsRef<str>>(&self, typepath: S) -> bool

Checks whether this Value’s type is equal to typepath.

pub fn is_truthy(&self) -> bool

pub fn from_string<S: AsRef<str>>(data: S) -> DMResult

Creates a Value that references a byond string. Will panic if the given string contains null bytes

Examples:

let my_string = Value::from_string("Testing!");

pub fn from_string_raw(data: &[u8]) -> DMResult

pub unsafe fn from_raw(v: Value) -> Self

blah blah lifetime is not verified with this so use at your peril

pub unsafe fn from_raw_owned(v: Value) -> Value

same as from_raw but does not increment the reference count (assumes we already own this reference)

impl Value

pub fn as_weak(&self) -> DMResult<WeakValue>

Creates a WeakValue referencing this datum.